LFRS attack known plain-cipher pair and degree
Self-calculate quite simple, my natural way is drawing a table with p0,..,pm row and states column, then put the known key on (which you already calculate it out from plain-cipher pair) then calculate follow the main diagonal of every sub matrix. That's my way, I dont' know how it calls.
Btw, the condition of this breakable case is the size of plain-cipher pair. In my opinion, the size should be large than max(m-p+1,p) in which p is period of key, m is degree of LFRS systems.