Substitution cipher attack by frequency using SD
My idea is using SD (Standard Deviation) to, you know, bounding the possible case of substitution. Since a letter be a binomial, for instance 12% it can be E or 78% not, it should has SD of binomial distribution.
Actually, after running this code, you'll have a 'mapping table', which show all possible case of substitution, all you wanna do is sort it decrease by number of case then trying to find the answer. Goodluck!